+++ include/lifobuffer.h
@@ -2,19 +2,19 @@
 #define _LIFO_BUFFER_H_
 
 #include <stdbool.h>
-#include <stdint.h>
+#include <stddef.h>
 
 typedef struct
 {
 	void*	 Buffer;
-	uint32_t ElementSize;
-	uint32_t NumElements;
-	uint32_t Head;
+	size_t ElementSize;
+	size_t NumElements;
+	size_t Head;
 	bool	 Full;
 	bool	 Empty;
 } LifoBuffer;
 
-void  LifoBuffer_Init(LifoBuffer* lifo, void* buff, uint32_t elementSize, uint32_t numElements);
+void  LifoBuffer_Init(LifoBuffer* lifo, void* buff, size_t elementSize, size_t numElements);
 void* LifoBuffer_Add(LifoBuffer* lifo);
 void* LifoBuffer_Remove(LifoBuffer* lifo);
